Add EngineRequest class
