+++ engines/ahmia/hidden_service.php
@@ -27,27 +27,4 @@
 
         return $results;
     }
-
-    function print_hidden_service_results($results)
-    {
-        echo "<div class=\"text-result-container\">";
-
-        foreach($results as $result)
-        {
-            $title = $result["title"];
-            $url = $result["url"];
-            $base_url = $result["base_url"];
-            $description = $result["description"];
-
-            echo "<div class=\"text-result-wrapper\">";
-            echo "<a href=\"$url\">";
-            echo "$base_url";
-            echo "<h2>$title</h2>";
-            echo "</a>";
-            echo "<span>$description</span>";
-            echo "</div>";
-        }
-
-        echo "</div>";
-    }
 ?>
