+++ engines/special/ip.php
@@ -1,11 +1,17 @@
 <?php
-    function ip_result()
-    {
-            return array(
-                "special_response" => array(
-                    "response" => $_SERVER["REMOTE_ADDR"],
-                    "source" => null
-                )
-            );
+    class IPRequest extends EngineRequest {
+        function __construct($query, $mh, $config) {
+            $this->query = $query;
+        }
+
+        function get_results()
+        {
+                return array(
+                    "special_response" => array(
+                        "response" => $_SERVER["REMOTE_ADDR"],
+                        "source" => null
+                    )
+                );
+        }
     }
 ?>
