+++ engines/special/user_agent.php
@@ -1,11 +1,17 @@
 <?php
-    function user_agent_result()
-    {
+    class UserAgentRequest extends EngineRequest {
+        function __construct($query, $mh, $config) {
+            $this->query = $query;
+        }
+
+        function get_results()
+        {
             return array(
                 "special_response" => array(
                     "response" => $_SERVER["HTTP_USER_AGENT"], 
                     "source" => null
                 )
             );                   
+        }
     }
 ?>
