+++ engines/special/weather.php
@@ -1,26 +1,31 @@
 <?php
-    function weather_results($response)
-    {
+    class WeatherRequest extends EngineRequest {
+        public function get_request_url () {
+            return "https://wttr.in/@" . $_SERVER["REMOTE_ADDR"] . "?format=j1";
+        }
+        
+        public function get_results() {
+            $response = curl_multi_getcontent($this->ch);
             $json_response = json_decode($response, true);
 
-            if ($json_response)
-            {
-                $current_weather = $json_response["current_condition"][0];
+            if (!$json_response)
+                return array();
 
-                $temp_c = $current_weather["temp_C"];
-                $temp_f = $current_weather["temp_F"];
-                $description = $current_weather["weatherDesc"][0]["value"];
+            $current_weather = $json_response["current_condition"][0];
 
-                $formatted_response = "$description - $temp_c °C | $temp_f °F";
+            $temp_c = $current_weather["temp_C"];
+            $temp_f = $current_weather["temp_F"];
+            $description = $current_weather["weatherDesc"][0]["value"];
 
-                $source = "https://wttr.in";
-                return array(
-                    "special_response" => array(
-                        "response" => htmlspecialchars($formatted_response),
-                        "source" => $source
-                    )
-                );
-            }
+            $formatted_response = "$description - $temp_c °C | $temp_f °F";
 
+            $source = "https://wttr.in";
+            return array(
+                "special_response" => array(
+                    "response" => htmlspecialchars($formatted_response),
+                    "source" => $source
+                )
+            );
+        }
     }
 ?>
