+++ engines/special/wikipedia.php
@@ -1,21 +1,27 @@
 <?php
-    function wikipedia_results($query, $response) 
-    {
-        global $config;
+    class WikipediaRequest extends EngineRequest {
+        public function get_request_url() {
+                $this->wikipedia_language = isset($_COOKIE["wikipedia_language"]) ? trim(htmlspecialchars($_COOKIE["wikipedia_language"])) : $this->config->wikipedia_language;
+                $query_encoded = urlencode($this->query);
 
-        $query_encoded = urlencode($query);
+                if (in_array($this->wikipedia_language, json_decode(file_get_contents("static/misc/wikipedia_langs.json"), true)))
+                    return "https://$this->wikipedia_language.wikipedia.org/w/api.php?format=json&action=query&prop=extracts%7Cpageimages&exintro&explaintext&redirects=1&pithumbsize=500&titles=$query_encoded";
 
-        $json_response = json_decode($response, true);
+                return "";
+        }
 
-        $first_page = array_values($json_response["query"]["pages"])[0];
+        public function get_results() {
+            $response = curl_multi_getcontent($this->ch);
 
-        if (!array_key_exists("missing", $first_page))
-        {
+            $json_response = json_decode($response, true);
+
+            $first_page = array_values($json_response["query"]["pages"])[0];
+
+            if (array_key_exists("missing", $first_page))
+                return array();
             $description = substr($first_page["extract"], 0, 250) . "...";
 
-            $wikipedia_language = isset($_COOKIE["wikipedia_language"]) ? trim(htmlspecialchars($_COOKIE["wikipedia_language"])) : $config->wikipedia_language;
-
-            $source = check_for_privacy_frontend("https://$wikipedia_language.wikipedia.org/wiki/$query");
+            $source = check_for_privacy_frontend("https://$wikipedia_language.wikipedia.org/wiki/$this->query");
             $response = array(
                 "special_response" => array(
                     "response" => htmlspecialchars($description),
