+++ misc/tools.php
@@ -101,86 +101,6 @@
         return $url;
     }
 
-    function check_ddg_bang($query)
-    {
-
-        $bangs_json = file_get_contents("static/misc/ddg_bang.json");
-        $bangs = json_decode($bangs_json, true);
-
-        if (substr($query, 0, 1) == "!")
-            $search_word = substr(explode(" ", $query)[0], 1);
-        else
-            $search_word = substr(end(explode(" ", $query)), 1);
-        
-        $bang_url = null;
-
-        foreach($bangs as $bang)
-        {
-            if ($bang["t"] == $search_word)
-            {
-                $bang_url = $bang["u"];
-                break;
-            }
-        }
-
-        if ($bang_url)
-        {
-            $bang_query_array = explode("!" . $search_word, $query);
-            $bang_query = trim(implode("", $bang_query_array));
-
-            $request_url = str_replace("{{{s}}}", str_replace('%26quot%3B','%22', urlencode($bang_query)), $bang_url);
-            $request_url = check_for_privacy_frontend($request_url);
-
-            header("Location: " . $request_url);
-            die();
-        }
-    }
-
-    function check_for_special_search($query)
-    {
-        if (isset($_COOKIE["disable_special"]))
-            return 0;
-
-         $query_lower = strtolower($query);
-         $split_query = explode(" ", $query);
-
-         if (strpos($query_lower, "to") && count($split_query) >= 4) // currency
-         {
-            $amount_to_convert = floatval($split_query[0]);
-            if ($amount_to_convert != 0)
-                return 1;
-         }
-         else if (strpos($query_lower, "mean") && count($split_query) >= 2) // definition
-         {
-             return 2;
-         }
-         else if (strpos($query_lower, "my") !== false)
-         {
-            if (strpos($query_lower, "ip"))
-            {
-                return 3;
-            }
-            else if (strpos($query_lower, "user agent") || strpos($query_lower, "ua"))
-            {
-                return 4;
-            }
-         }
-         else if (strpos($query_lower, "weather") !== false)
-         {
-                return 5;
-         }
-         else if ($query_lower == "tor")
-         {
-                return 6;
-         }
-         else if (3 > count(explode(" ", $query))) // wikipedia
-         {
-             return 7;
-         }
-
-        return 0;
-    }
-
     function get_xpath($response)
     {
         $htmlDom = new DOMDocument;
